feat: supervisor

+++ run_server.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 # Xiang Wang(ramwin@qq.com)
 
-rm -r _build
-sphinx-autobuild -j auto --port 18001 . _build/html/ --re-ignore "\.git"
+if [ -d _build ]
+then
+    rm -r _build
+fi
+
+sphinx-autobuild \
+    -j auto \
+    --port 18001 \
+    . _build/html/ \
+    --re-ignore "\.mypy_cache" \
+    --re-ignore "\.git"    \
+    --re-ignore "\.*\.swp"
